Backend/requirements.txt
Flask
psycopg2-binary
Flask-cors